Outbox Improvements for Relation Database (#3464)
* Fixes MySQL queries for Dispatched Commands

* Fixes MySQL Outbox queries

* Fixes Postgres outbox

* Fix Sqlite outbox

* Fixes MsSQL outbox

* Fixes compilation issue

* Fixes MSSQL when convert the message id

+++ src/Paramore.Brighter.MessagingGateway.MsSql/SqlQueues/MsSqlMessageQueue.cs
@@ -157,7 +157,7 @@ namespace Paramore.Brighter.MessagingGateway.MsSql.SqlQueues
                 return ReceivedResult<T>.Empty;
             var json = (string) reader[0];
             var messageType = (string) reader[1];
-            var id = (int) reader[3];
+            var id = (long) reader[3];
             var message = JsonSerializer.Deserialize<T>(json, JsonSerialisationOptions.Options);
             return new ReceivedResult<T>(true, json, topic, messageType, id, message);
         }
